requirements_dev.txt
black==24.8.0
flake8==7.1.1
mongomock-motor==0.0.34
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
setuptools==75.8.0
mypy==1.14.1
tox==4.23.2


discord.py==2.4.0
python-dotenv==1.0.1
pytz==2024.2
mongoengine==0.29.1
motor==3.6.0
Flask==3.1.0
requests==2.32.3
